+++ hero-be/src/main/java/com/c4/hero/domain/employee/controller/EmployeeController.java
@@ -65,9 +65,11 @@ import java.util.List;
 public class EmployeeController {
 
     private final EmployeeCommandService employeeCommandService;
+    private final EmployeePasswordService employeePasswordService;
     private final EmployeeQueryService employeeQueryService;
     private final EmployeeProfileQueryService employeeProfileQueryService;
     private final EmployeeSealService employeeSealService;
+
 
 
     private final JwtUtil jwtUtil;
